allow constraints created/dropped by kind of constraint
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@5553 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ dsmf/src/com/exedio/dsmf/Table.java
@@ -327,29 +327,29 @@ public final class Table extends Node
 
 	}
 	
-	final void createConstraints(final boolean secondPhase)
+	final void createConstraints(final int mask, final boolean secondPhase)
 	{
 		for(final Constraint constraint : constraintList)
 		{
-			if(constraint.secondPhase==secondPhase)
+			if(constraint.matchesMask(mask) && constraint.secondPhase==secondPhase)
 				constraint.create();
 		}
 	}
 	
-	final void dropConstraints(final boolean secondPhase)
+	final void dropConstraints(final int mask, final boolean secondPhase)
 	{
 		for(final Constraint constraint : constraintList)
 		{
-			if(constraint.secondPhase==secondPhase)
+			if(constraint.matchesMask(mask) && constraint.secondPhase==secondPhase)
 				constraint.drop();
 		}
 	}
 	
-	final void tearDownConstraints(final boolean secondPhase)
+	final void tearDownConstraints(final int mask, final boolean secondPhase)
 	{
 		for(final Constraint constraint : constraintList)
 		{
-			if(constraint.secondPhase==secondPhase)
+			if(constraint.matchesMask(mask) && constraint.secondPhase==secondPhase)
 			{
 				try
 				{
